Completed regex-based blacklist checking
The message handler has been updated to check every message against a json blacklist file containing "blacklist string":"moderation action" pairings. The blacklist has
simple regex support (handled by placing * at the beginning or end of a string) to allow for checking for the blacklist string as any part of a message, as the
beginning of a word, as the end of a word, or as an entire word (Note: word* and *word will also match the word by itself.