Refactored utility functions and structs and began blacklist command implementation
Due to common use across multiple modules, formatOutgoingMessage(), stringToMessagE(), saveConfig() (which has been renamed saveJSON and made general purpose for use
with the blacklist JSONValue), and the Message struct have been migrated to a new file: spambot_util.d

Command parsing now does not break when the command is a single term (instead it checks for the presence of a space before isolating the command string). The blacklist
command parser checks that the command is at least 3 terms before parsing and sends an error to stderr if the check fails. Command term isolation now works for
blacklist. A comment block has been added outlining the names and features of planned blacklist subcommands.